Refactor ConnectWalletModal and JoinCampaignOverlay to use custom hook for wallet options management
